Move deprecated totals rollups into quantum task
