+++ lib/castle/application.ex
@@ -10,8 +10,7 @@ defmodule Castle.Application do
     children = [
       supervisor(Castle.Repo, []),
       supervisor(CastleWeb.Endpoint, []),
-      worker(Castle.Scheduler, []),
-      worker(Castle.Rollup.Worker, [])
+      worker(Castle.Scheduler, [])
     ]
 
     # Create the redix children list of workers:
